+++ libs/base_core/src/text/char/namespace.rs
@@ -5,6 +5,7 @@
 // TOC
 // - methods over char
 // - methods over u32
+// - methods over u16
 // - methods over u8
 // - methods over &[u8]
 
@@ -17,6 +18,7 @@ use crate::{ASCII_TABLE, unwrap};
 /// # Methods
 /// - [over `char`](#methods-over-char)
 /// - [over `u32`](#methods-over-u32)
+/// - [over `u16`](#methods-over-u16)
 /// - [over `u8`](#methods-over-u8)
 /// - [over `&[u8]`](#methods-over-u8-slice)
 ///
@@ -174,6 +176,24 @@ impl Char<u32> {
             || (self.0 >= 0x2FE0 && self.0 <= 0x2FEF)
     }
 
+    /// Returns `true` if the given unicode scalar code is a [surrogate code point][0].
+    ///
+    /// [0]: https://www.unicode.org/glossary/#surrogate_code_point
+    #[must_use] #[inline(always)]
+    pub const fn is_surrogate(self) -> bool { matches!(self.0, 0xD800..=0xDFFF) }
+
+    /// Returns `true` if the given unicode scalar code is a [leading surrogate][0].
+    ///
+    /// [0]: https://www.unicode.org/glossary/#high_surrogate_code_point
+    #[must_use] #[inline(always)]
+    pub const fn is_surrogate_high(self) -> bool { matches!(self.0, 0xD800..=0xDBFF) }
+
+    /// Returns `true` if the given unicode scalar code is a [trailing surrogate][0].
+    ///
+    /// [0]: https://www.unicode.org/glossary/#low_surrogate_code_point
+    #[must_use] #[inline(always)]
+    pub const fn is_surrogate_low(self) -> bool { matches!(self.0, 0xDC00..=0xDFFF) }
+
     /// Returns the ASCII representation as a `&'static str`, or `""` if non-ASCII.
     #[must_use]
     pub const fn code_to_ascii_str(self) -> &'static str {
@@ -238,6 +258,45 @@ impl Char<u32> {
                 let x = 0b11110_000 | ((code >> 18) as u8);
                 [x, y, z, w]
             }
+        }
+    }
+}
+
+/// # Methods over `u16`.
+#[rustfmt::skip]
+impl Char<u16> {
+    #[must_use] #[inline(always)]
+    /// Returns `true` if the given unicode scalar code is a [surrogate code point][0].
+    ///
+    /// [0]: https://www.unicode.org/glossary/#surrogate_code_point
+    pub const fn is_surrogate(self) -> bool { matches!(self.0, 0xD800..=0xDFFF) }
+
+    #[must_use] #[inline(always)]
+    /// Returns `true` if the given unicode scalar code is a [leading surrogate][0].
+    ///
+    /// [0]: https://www.unicode.org/glossary/#high_surrogate_code_point
+    pub const fn is_surrogate_high(self) -> bool { matches!(self.0, 0xD800..=0xDBFF) }
+
+    #[must_use] #[inline(always)]
+    /// Returns `true` if the given unicode scalar code is a [trailing surrogate][0].
+    ///
+    /// [0]: https://www.unicode.org/glossary/#low_surrogate_code_point
+    pub const fn is_surrogate_low(self) -> bool { matches!(self.0, 0xDC00..=0xDFFF) }
+
+    #[must_use]
+    /// Decodes the given surrogate pair.
+    ///
+    /// # Features
+    /// Makes use of the `unsafe_str` feature if enabled to avoid double validation.
+    pub const fn decode_surrogate_pair(high: u16, low: u16) -> Option<char> {
+        if Char(high).is_surrogate_high() && Char(low).is_surrogate_low() {
+            let code = 0x10000 + (((high as u32 - 0xD800) << 10) | (low as u32 - 0xDC00));
+            #[cfg(any(base_safe_text, not(feature = "unsafe_str")))]
+            return char::from_u32(code);
+            #[cfg(all(not(base_safe_text), feature = "unsafe_str"))]
+            Some(unsafe { char::from_u32_unchecked(code) })
+        } else {
+            None
         }
     }
 }
@@ -364,9 +423,9 @@ impl Char<&[u8]> {
 /// Methods over a byte array, referring to a byte slice.
 impl<const N: usize> Char<&[u8; N]> {
     /// See [utf8_bytes_to_code()][Char::<&[u8]>::utf8_bytes_to_code].
+    #[inline(always)] // eliminate the wrapper entirely
     pub const fn utf8_bytes_to_code(self, index: usize) -> Option<(u32, usize)> {
         let bytes: &[u8] = self.0;
-        // reuse the same implementation by converting to slice
         Char(bytes).utf8_bytes_to_code(index)
     }
 }